Added font fallback support
Font fallback should be done when a character isn't found in the series
of fonts already provided. At this point we should query for a font
based on the character set and not on what the user provided.
Optionally we can also pass a pattern to make it more or less close to
what the user initially queried for (monospace, italic, etc..).